requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiogram3-calendar==0.1.2
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
YooMoney==0.1.0
